+++ Bukkit/src/main/java/info/preva1l/fadah/data/handler/MySQLHandler.java
@@ -125,11 +125,6 @@ public class MySQLHandler implements DatabaseHandler {
     }
 
     @Override
-    public void wipeDatabase() {
-        // nothing yet
-    }
-
-    @Override
     public <T> List<T> getAll(Class<T> clazz) {
         return (List<T>) getDao(clazz).getAll();
     }
