+++ Bukkit/src/main/java/info/preva1l/fadah/guis/CollectionBoxMenu.java
@@ -29,11 +29,7 @@ public class CollectionBoxMenu extends PaginatedFastInv {
         this.viewer = viewer;
         this.owner = owner;
 
-        List<Integer> fillerSlots = getLayout().fillerSlots();
-        if (!fillerSlots.isEmpty()) {
-            setItems(fillerSlots.stream().mapToInt(Integer::intValue).toArray(),
-                    GuiHelper.constructButton(GuiButtonType.BORDER));
-        }
+        fillers();
         setPaginationMappings(getLayout().paginationSlots());
 
         addNavigationButtons();
@@ -69,24 +65,6 @@ public class CollectionBoxMenu extends PaginatedFastInv {
                     CacheAccess.getNotNull(History.class, owner.getUniqueId()).add(historicItem);
                 }, null, 0L);
             }));
-        }
-    }
-
-    @Override
-    protected void addPaginationControls() {
-        setItem(getLayout().buttonSlots().getOrDefault(LayoutManager.ButtonType.PAGINATION_CONTROL_ONE, -1),
-                GuiHelper.constructButton(GuiButtonType.BORDER));
-        setItem(getLayout().buttonSlots().getOrDefault(LayoutManager.ButtonType.PAGINATION_CONTROL_TWO,-1),
-                GuiHelper.constructButton(GuiButtonType.BORDER));
-        if (page > 0) {
-            setItem(getLayout().buttonSlots().getOrDefault(LayoutManager.ButtonType.PAGINATION_CONTROL_ONE, -1),
-                    GuiHelper.constructButton(GuiButtonType.PREVIOUS_PAGE), e -> previousPage());
-        }
-
-        var collectionBox = CacheAccess.getNotNull(CollectionBox.class, owner.getUniqueId()).collectableItems();
-        if (collectionBox != null && collectionBox.size() >= index + 1) {
-            setItem(getLayout().buttonSlots().getOrDefault(LayoutManager.ButtonType.PAGINATION_CONTROL_TWO,-1),
-                    GuiHelper.constructButton(GuiButtonType.NEXT_PAGE), e -> nextPage());
         }
     }
