+++ Bukkit/src/main/java/info/preva1l/fadah/guis/ExpiredListingsMenu.java
@@ -30,13 +30,8 @@ public class ExpiredListingsMenu extends PaginatedFastInv {
         this.viewer = viewer;
         this.owner = owner;
 
-        List<Integer> fillerSlots = getLayout().fillerSlots();
-        if (!fillerSlots.isEmpty()) {
-            setItems(fillerSlots.stream().mapToInt(Integer::intValue).toArray(),
-                    GuiHelper.constructButton(GuiButtonType.BORDER));
-        }
+        fillers();
         setPaginationMappings(getLayout().paginationSlots());
-
         addNavigationButtons();
         fillPaginationItems();
         populatePage();
@@ -69,24 +64,6 @@ public class ExpiredListingsMenu extends PaginatedFastInv {
                     CacheAccess.getNotNull(History.class, owner.getUniqueId()).add(historicItem);
                 },null, 0L);
             }));
-        }
-    }
-
-    @Override
-    protected void addPaginationControls() {
-        setItem(getLayout().buttonSlots().getOrDefault(LayoutManager.ButtonType.PAGINATION_CONTROL_ONE, -1),
-                GuiHelper.constructButton(GuiButtonType.BORDER));
-        setItem(getLayout().buttonSlots().getOrDefault(LayoutManager.ButtonType.PAGINATION_CONTROL_TWO,-1),
-                GuiHelper.constructButton(GuiButtonType.BORDER));
-        if (page > 0) {
-            setItem(getLayout().buttonSlots().getOrDefault(LayoutManager.ButtonType.PAGINATION_CONTROL_ONE, -1),
-                    GuiHelper.constructButton(GuiButtonType.PREVIOUS_PAGE), e -> previousPage());
-        }
-
-        var expiredItems = CacheAccess.getNotNull(ExpiredItems.class, owner.getUniqueId()).expiredItems();
-        if (expiredItems != null && expiredItems.size() >= index + 1) {
-            setItem(getLayout().buttonSlots().getOrDefault(LayoutManager.ButtonType.PAGINATION_CONTROL_TWO,-1),
-                    GuiHelper.constructButton(GuiButtonType.NEXT_PAGE), e -> nextPage());
         }
     }
